# swagger-spec/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/regions": {
      "post": {
        "operationId": "RegionController_createRegion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah Region Baru (Superadmin Only)",
        "tags": [
          "Wilayah & Kota"
        ]
      },
      "get": {
        "operationId": "RegionController_findAllRegions",
        "parameters": [
          {
            "name": "onlyActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Melihat seluruh daftar Region beserta Admin Penanggung Jawab",
        "tags": [
          "Wilayah & Kota"
        ]
      }
    },
    "/api/regions/{id}": {
      "get": {
        "operationId": "RegionController_findOneRegion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Region ditemukan"
          },
          "404": {
            "description": "Region tidak ditemukan"
          }
        },
        "summary": "Melihat detail Region & daftar Admin Wilayah",
        "tags": [
          "Wilayah & Kota"
        ]
      },
      "patch": {
        "operationId": "RegionController_updateRegion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Region berhasil diperbarui"
          },
          "404": {
            "description": "Region tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update/Deaktivasi Region (Superadmin Only)",
        "tags": [
          "Wilayah & Kota"
        ]
      }
    },
    "/api/cities": {
      "post": {
        "operationId": "RegionController_createCity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kota berhasil dibuat"
          },
          "409": {
            "description": "Kota sudah terdaftar"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah Kota Baru (Superadmin & Admin Wilayah)",
        "tags": [
          "Wilayah & Kota"
        ]
      },
      "get": {
        "operationId": "RegionController_findAllCities",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Melihat seluruh daftar Kota",
        "tags": [
          "Wilayah & Kota"
        ]
      }
    },
    "/api/pickup-points": {
      "post": {
        "operationId": "PickupPointController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePickupPointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pos resmi berhasil dibuat"
          },
          "400": {
            "description": "Payload atau role operator tidak valid"
          },
          "404": {
            "description": "Region, City, atau Operator tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah Pickup Point / Pos Resmi Baru (Admin/Superadmin)",
        "tags": [
          "Pickup Points (Pos Resmi)"
        ]
      },
      "get": {
        "operationId": "PickupPointController_findAll",
        "parameters": [
          {
            "name": "regionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "onlyActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar Pos Resmi ditemukan"
          }
        },
        "summary": "Melihat seluruh Pos Resmi",
        "tags": [
          "Pickup Points (Pos Resmi)"
        ]
      }
    },
    "/api/pickup-points/{id}": {
      "get": {
        "operationId": "PickupPointController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pos resmi ditemukan"
          },
          "404": {
            "description": "Pos Resmi tidak ditemukan"
          }
        },
        "summary": "Melihat detail Pos Resmi berdasarkan ID",
        "tags": [
          "Pickup Points (Pos Resmi)"
        ]
      },
      "patch": {
        "operationId": "PickupPointController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePickupPointDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pos resmi berhasil diperbarui"
          },
          "403": {
            "description": "Bukan pos resmi wilayah Anda"
          },
          "404": {
            "description": "Pos resmi tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update/Deaktivasi Pos Resmi (Admin/Superadmin)",
        "tags": [
          "Pickup Points (Pos Resmi)"
        ]
      }
    },
...
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pesanan berhasil dibuat"
          },
          "400": {
            "description": "Sisa kursi/bagasi tidak mencukupi atau mencoba memesan di trip sendiri"
          },
          "404": {
            "description": "Jadwal trip tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buat pesanan baru (Passenger / Parcel Booking)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/me": {
      "get": {
        "operationId": "OrdersController_getMyOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar pesanan ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Daftar riwayat pesanan milik Customer yang sedang login",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail pesanan ditemukan"
          },
          "403": {
            "description": "Tidak memiliki hak akses ke order ini"
          },
          "404": {
            "description": "Order tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detail pesanan berdasarkan ID (Dengan proteksi otorisasi)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrdersController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pesanan berhasil dibatalkan"
          },
          "400": {
            "description": "Pesanan sudah selesai/berjalan atau tidak dapat dibatalkan"
          },
          "403": {
            "description": "Tidak memiliki hak akses untuk membatalkan pesanan ini"
          },
          "404": {
            "description": "Order tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Membatalkan pesanan dan mengembalikan kuota kursi/bagasi trip",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nebeng App API Specification",
    "description": "Dokumentasi API Backend Platform Nebeng Transportasi & Pengiriman Paket (Escrow Wallet & Dual QR Checkpoint)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Masukkan token JWT: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateRegionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Wilayah Yogyakarta",
            "description": "nama Wilayah"
          },
          "code": {
            "type": "string",
            "example": "REG-DIY",
            "description": "Kode unik wilayah"
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
      "UpdateRegionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Region Yogyakarta Updated"
          },
          "code": {
            "type": "string",
            "example": "REG-DIY"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "CreateCityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sleman",
            "description": "Nama kota/kabupaten"
          },
          "province": {
            "type": "string",
            "example": "DI Yogyakarta",
            "description": "Nama Provinsi"
          }
        },
        "required": [
          "name",
          "province"
        ]
      },
      "CreatePickupPointDto": {
        "type": "object",
        "properties": {
          "regionId": {
            "type": "string",
            "example": "1",
            "description": "ID Region"
          },
          "cityId": {
            "type": "string",
            "example": "1",
            "description": "ID City"
          },
          "operatorId": {
            "type": "string",
            "example": "2",
            "description": "ID User Operator Pos"
          },
          "name": {
            "type": "string",
            "example": "Pos Terminal Giwangan",
            "description": "Nama Pos"
          },
          "address": {
            "type": "string",
            "example": "Jl. Imogiri Timur No.1, Yogyakarta",
            "description": "Alamat lengkap"
          },
          "latitude": {
            "type": "number",
            "example": -7.8333,
            "description": "Latitude"
          },
          "longitude": {
            "type": "number",
            "example": 110.3833,
            "description": "Longitude"
          }
        },
        "required": [
          "regionId",
          "cityId",
          "name",
          "address",
          "latitude",
          "longitude"
        ]
      },
      "UpdatePickupPointDto": {
        "type": "object",
        "properties": {
          "regionId": {
            "type": "string",
            "example": "1"
          },
          "cityId": {
            "type": "string",
            "example": "1"
          },
          "operatorId": {
            "type": "string",
            "example": "2"
          },
          "name": {
            "type": "string",
            "example": "Pos Giwangan Baru"
          },
          "address": {
            "type": "string",
            "example": "Jl. Imogiri Timur No.10"
          },
          "latitude": {
            "type": "number",
            "example": -7.8333
          },
          "longitude": {
            "type": "number",
            "example": 110.3833
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string",
            "example": "Dokumen Penting"
          },
          "itemCategory": {
            "type": "string",
            "example": "Dokumen"
          },
          "quantity": {
            "type": "number",
            "example": 1
          },
          "weightPerItemKg": {
            "type": "number",
            "example": 2.5,
            "description": "Berat per item dalam KG"
          },
          "sizeEnum": {
            "type": "string",
            "enum": [
              "xxs",
              "xs",
              "s",
              "m",
              "l",
              "xl"
            ],
            "example": "s"
          },
          "photoUrl": {
            "type": "string",
            "example": "https://storage.nebeng.com/item1.jpg"
          },
          "recipientName": {
            "type": "string",
            "example": "Budi Santoso",
            "description": "Nama penerima paket"
          },
          "recipientPhone": {
            "type": "string",
            "example": "081234567890",
            "description": "Nomor HP penerima paket"
          }
        },
        "required": [
          "itemName",
          "itemCategory",
          "quantity",
          "weightPerItemKg",
          "sizeEnum",
          "recipientName",
          "recipientPhone"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "example": "1",
            "description": "ID Trip yang dipesan"
          },
          "type": {
            "type": "string",
            "enum": [
              "passenger",
              "parcel"
            ],
            "example": "passenger"
          },
          "seatsBooked": {
            "type": "number",
            "example": 1,
            "description": "Jumlah kursi (Wajib diisi jika type = passenger)"
          },
          "items": {
            "description": "Daftar item paket (Wajib diisi jika type = parcel)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "tripId",
          "type"
        ]
      },
...
    }
  }
}